added some timer stuffs to make life easier

+++ src/wd_debug.h
@@ -1,17 +1,20 @@
 /**
+ * @file
+ * 
+ * @brief      Debug utils
+ *
  * @addtogroup WD
  * @{
  * 
  * @addtogroup STDIO
  * @{
- * 
- * @brief      Debug utils
  */
 
 #pragma once
 
 #include "wd_dev_uart1.h"
 #include "../include/config.h"
+#include <stdint.h>
 
 extern void wd_panic;
 
@@ -22,10 +25,10 @@ extern void wd_panic;
 #define __UART1_DEBUG_PRINTF(format, ...)
 #endif
 
-#define WD_INFO(format, ...)  __UART1_DEBUG_PRINTF("[WD\tINFO\t%s:%d] " format, __FILE__, __LINE__, ##__VA_ARGS__)
-#define WD_WARN(format, ...)  __UART1_DEBUG_PRINTF("[WD\tWARN\t%s:%d] " format, __FILE__, __LINE__, ##__VA_ARGS__)
-#define WD_ERROR_SOFT(format, ...) __UART1_DEBUG_PRINTF("[WD\tERROR\t%s:%d] " format, __FILE__, __LINE__, ##__VA_ARGS__);
-#define WD_ERROR(format, ...) WD_ERROR_SOFT(format, ##__VA_ARGS__); goto wd_panic
+#define WD_INFO(format, ...)       __UART1_DEBUG_PRINTF("[WD\tINFO\t%s:%d] " format, __FILE__, (uintmax_t) __LINE__, ##__VA_ARGS__)
+#define WD_WARN(format, ...)       __UART1_DEBUG_PRINTF("[WD\tWARN\t%s:%d] " format, __FILE__, (uintmax_t) __LINE__, ##__VA_ARGS__)
+#define WD_ERROR_SOFT(format, ...) __UART1_DEBUG_PRINTF("[WD\tERROR\t%s:%d] " format, __FILE__, (uintmax_t) __LINE__, ##__VA_ARGS__);
+#define WD_ERROR(format, ...)      WD_ERROR_SOFT(format, ##__VA_ARGS__); goto wd_panic
 
 #define WD_ASSERT_SOFT(condition, then) \
     if (!(condition)) { \
